requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
plotly==5.17.0
matplotlib==3.7.2
openpyxl==3.1.2
reportlab==4.0.5
Pillow==10.0.0
scipy==1.11.2
python-dotenv==1.0.0
python-docx==1.1.0
telethon==1.31.0
nest-asyncio==1.5.8
